Discontinuing
Rationale: The Habitica Sync plugin exists, and while although the
implementation instead chooses to use an Obsidian Pane and is not
capable of integrating with Obsidian Tasks, I've realized that doing
so honestly introduces it's own set of problems, and is perhaps a tad
redundant and unnecessary. All of the desired features of the
Obsidian Tasks plugin is already available via Habitica.